w3c-reports/12092014/mac-wk2/runned-tests/html/semantics/forms/textfieldselection/selection-expected.txt



PASS test if selection text is correct for input 
PASS test if selection text is correct for textarea 
PASS test if non-ascii selection text is correct for input 
PASS test if non-ascii selection text is correct for textarea 
PASS test SelectionStart offset for input 
FAIL test SelectionStart offset for textarea assert_equals: SelectionStart offset without selection expected 10 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/selection.html:99:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/selection.html:95:7)
PASS test SelectionEnd offset for input 
FAIL test SelectionEnd offset for textarea assert_equals: SelectionEnd offset without selection expected 10 but got 0(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/html/semantics/forms/textfieldselection/selection.html:121:18
step@http://localhost:8800/resources/testharness.js:798:30
test@http://localhost:8800/resources/testharness.js:73:22
global code@http://localhost:8800/html/semantics/forms/textfieldselection/selection.html:117:7)
PASS test SelectionDirection for input 
PASS test SelectionDirection for textarea 
